+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__
 *.mp4
 
 resources/
+temp/
