+++ aruco_markerdetection/aruco.py
@@ -21,28 +21,28 @@ def detect(frame, camera_matrix, dist_coefs, obj):
     corners, ids, rejectedImgPoints = aruco.detectMarkers(frame_gray, aruco_dict, parameters=parameters)
     # frame = aruco.drawDetectedMarkers(frame, corners, ids)
 
-    src_pts = np.float32([ [0,0],[0,h-1],[w-1,0],[w-1,h-1] ]).reshape(-1,1,2)
+    # src_pts = np.float32([ [0,0],[0,h-1],[w-1,0],[w-1,h-1] ]).reshape(-1,1,2)
     rendered_img = frame
-    rvec = []
-    tvec = []
+    rvecs = []
+    tvecs = []
     projection_m = []
     center = []
     if corners != []:
-        rvecs, tvecs, _ = aruco.estimatePoseSingleMarkers(corners, 0.05, camera_matrix, dist_coefs)
+        rvecs, tvecs, _ = aruco.estimatePoseSingleMarkers(corners, 8.0, camera_matrix, dist_coefs)
         rvec = rvecs[0]
         tvec = tvecs[0]
-        rendered_img = aruco.drawAxis(rendered_img, camera_matrix, dist_coefs, rvec, tvec, 0.03)
+        rendered_img = aruco.drawAxis(rendered_img, camera_matrix, dist_coefs, rvec, tvec, 6.0)
 
-        c = corners[0][0]
-        center_x = c[:, 0].mean()
-        center_y = c[:, 1].mean()
-        center = [center_x, center_y]
+        # c = corners[0][0]
+        # center_x = c[:, 0].mean()
+        # center_y = c[:, 1].mean()
+        # center = [center_x, center_y]
 
         # rendered_img = cv2.circle(rendered_img,center,5,(255,0,0),20)
-        dst_pts = np.array(corners, dtype="float32").reshape(-1,1,2)
-        M, mask = cv2.findHomography(src_pts, dst_pts)
-        projection_m = projection_matrix(camera_matrix,M)
+        # dst_pts = np.array(corners, dtype="float32").reshape(-1,1,2)
+        # M, mask = cv2.findHomography(src_pts, dst_pts)
+        # projection_m = projection_matrix(camera_matrix,M)
         # _2d_points = show_axis(rendered_img,projection_m)
         # rendered_img = renderCube(rendered_img, rvec, tvec, camera_matrix, dist_coefs)
 
-    return (rvec, tvec, rendered_img, center, projection_m)
+    return (rvecs, tvecs, rendered_img, center, projection_m)
