+++ glut_application.py
@@ -1,189 +1,163 @@
 from OpenGL.GL import *
-from OpenGL.GLUT import *
 from OpenGL.GLU import *
-
+from OpenGL.GLUT import *
 import cv2
-
-from objloader_complete import *
-from utils.paths import *
+import numpy as np
+import time
 from utils.calibration import *
-from ppm_formula import *
-
-from PIL import Image
+from utils.objloader_complete import *
+from utils.paths import *
+from utils.webcam import *
+from aruco_markerdetection.aruco import *
 
 import ctypes
 
-from aruco import *
-
-###############################
-# GLOBAL VARIABLES
-width = 1280
-height = 720
 texture_background = None
-chess_piece = None
-webcam = cv2.VideoCapture(0)
+
+# Set AR
+aruco = cv2.aruco
+dictionary = aruco.Dictionary_get(aruco.DICT_6X6_250)
+
+#load camera parameter
 camera_matrix, dist_coefs, rvecs, tvecs = calibrate()
+webcam = Webcam(0)
 
-znear = 0.1
-zfar = 10000.0
+alpha = camera_matrix[0][0]
+beta = camera_matrix[1][1]
+cx = camera_matrix[0][2]
+cy = camera_matrix[1][2]
 
-###############################
+chess_piece = None
 
-############################### INIT OPENGL
-def init_gl():
-    global texture_background, chess_piece
+windowWidth = 1280
+windowHeight = 720
 
-    glClearColor(0.0, 0.0, 0.0, 0.0)
-    glClearDepth(1.0)
-    # glDepthFunc(GL_LESS)
-    glDepthFunc(GL_LEQUAL)
-    glEnable(GL_DEPTH_TEST)
-    glShadeModel(GL_SMOOTH)
+f = 1000.0
+n = 1.0
+
+def draw():
+
+    img = webcam.getNextFrame()
+
+    # Aruco
+    rvec, tvec, img, center, p_k = detect(img, camera_matrix, dist_coefs, chess_piece)
+    img = cv2.cvtColor(img,cv2.COLOR_BGR2RGB) #BGR-->RGB
+    h, w = img.shape[:2]
+
+    global texture_background
+    glBindTexture(GL_TEXTURE_2D, texture_background)
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, w, h, 0, GL_RGB, GL_UNSIGNED_BYTE, img)
+
+    ## Enable / Disable
+    glDisable(GL_DEPTH_TEST)    # Disable GL_DEPTH_TEST
+    glDisable(GL_LIGHTING)      # Disable Light
+    glDisable(GL_LIGHT0)        # Disable Light
+    glEnable(GL_TEXTURE_2D)     # Enable texture map
+
+    ## init
+    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT)  # Clear Buffer
+    glColor3f(1.0, 1.0, 1.0)    # Set texture Color(RGB: 0.0 ~ 1.0)
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR)
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR)
+
+    ## draw background
     glMatrixMode(GL_PROJECTION)
     glLoadIdentity()
-    gluPerspective(40.34, 1.78, znear, zfar)   #Cambiare all'occorrenza
-    glViewport(0,0,width,height)
-
-    # ppm = getOpenglProjectionMatrix(camera_matrix, znear, zfar, width, height)
-    # # glMatrixMode(GL_PROJECTION)
-    # # glPushMatrix()
-    # glLoadMatrixf(ppm)
-
-    glMatrixMode(GL_MODELVIEW)
-
-    # assign shapes
-    chess_piece = OBJ(obj_path, True)
-
-    # assign texture
-    glEnable(GL_TEXTURE_2D)
-    texture_background = glGenTextures(1)
-
-############################### WEBCAM HANDLER
-def getNextFrame():
-    global webcam
-
-    ret, frame = webcam.read()
-
-    if not ret or frame is None:
-        webcam.release()
-        print("Released webcam")
-        return None
-
-    return frame
-
-############################### WINDOWS FUNCTION
-def draw_background(image):
-    global texture_background
-
-    bg_image = cv2.flip(image, 0)
-    bg_image = Image.fromarray(bg_image)
-    ix = bg_image.size[0]
-    iy = bg_image.size[1]
-    bg_image = bg_image.tobytes("raw", "BGRX", 0, -1)
-
-    # create background texture
-    glBindTexture(GL_TEXTURE_2D, texture_background)
-    glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST)
-    glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST)
-    glTexImage2D(GL_TEXTURE_2D, 0, 3, ix, iy, 0, GL_RGBA, GL_UNSIGNED_BYTE, bg_image)
-
-    # draw background
     glBindTexture(GL_TEXTURE_2D, texture_background)
     glPushMatrix()
-    glTranslatef(0.0,0.0,-10.0)
-
     glBegin(GL_QUADS)
-    glTexCoord2f(0.0, 1.0); glVertex3f(-4.0, -3.0, 3.0)
-    glTexCoord2f(1.0, 1.0); glVertex3f( 4.0, -3.0, 3.0)
-    glTexCoord2f(1.0, 0.0); glVertex3f( 4.0,  3.0, 3.0)
-    glTexCoord2f(0.0, 0.0); glVertex3f(-4.0,  3.0, 3.0)
-    glEnd( )
-
+    glTexCoord2d(0.0, 1.0)
+    glVertex3d(-1.0, -1.0,  0)
+    glTexCoord2d(1.0, 1.0)
+    glVertex3d( 1.0, -1.0,  0)
+    glTexCoord2d(1.0, 0.0)
+    glVertex3d( 1.0,  1.0,  0)
+    glTexCoord2d(0.0, 0.0)
+    glVertex3d(-1.0,  1.0,  0)
+    glEnd()
     glPopMatrix()
 
-def draw_obj(rvec, tvec, center, p_k):
+    ## Enable / Disable
+    glEnable(GL_DEPTH_TEST)     # Enable GL_DEPTH_TEST
+    glEnable(GL_LIGHTING)       # Enable Light
+    glEnable(GL_LIGHT0)         # Enable Light
+    glDisable(GL_TEXTURE_2D)    # Disable texture map
 
-    ### OBTAINING PROJECTION MATRIX
-    view_matrix = getCorrectPPM(rvec,tvec)
-    # view_matrix = tryThis(p_k)
+    ## make projection matrix
+    m1 = np.array([
+    [(alpha)/cx, 0, 0, 0],
+    [0, beta/cy, 0, 0],
+    [0, 0, -(f+n)/(f-n), (-2.0*f*n)/(f-n)],
+    [0,0,-1,0],
+    ])
+    glLoadMatrixd(m1.T)
 
-    ###############################
-    lightfv = ctypes.c_float * 4
-    glDisable(GL_TEXTURE_2D)
-    glEnable(GL_DEPTH_TEST)
-    glLightfv(GL_LIGHT0, GL_POSITION, lightfv(-1.0, 1.0, 1.0, 0.0))
-    glEnable(GL_LIGHT0)
-
+    ## draw cube
     glMatrixMode(GL_MODELVIEW)
-    glPushMatrix()
-    glLoadMatrixd(view_matrix)
-    glTranslatef(0.0, 0.0, -100.0)
-    # glTranslatef(-center[0], center[1], 0.0)
-    # glRotatef(-90, 1.0, 0.0, 0.0)
-    glEnable(GL_LIGHTING)
-
-    glCallList(chess_piece.gl_list)
-
-    glColor3f(1.0, 1.0, 1.0)
-    glPopMatrix()
-    glEnable(GL_TEXTURE_2D)
-    glDisable(GL_DEPTH_TEST)
-    glDisable(GL_LIGHTING)
-
-
-def display():
-    global chess_piece, camera_matrix, dist_coefs
-
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT)
     glLoadIdentity()
+    glPushMatrix()  
 
-    image = getNextFrame()
+    # glMaterialfv(GL_FRONT_AND_BACK, GL_AMBIENT_AND_DIFFUSE, [0.0,0.0,1.0,1.0])
+    lightfv = ctypes.c_float * 4
+    glLightfv(GL_LIGHT0, GL_POSITION, lightfv(-1.0, 1.0, 1.0, 0.0))
+    if len(rvec) != 0:
+        # fix axis
+        tvec[0][0][0] = tvec[0][0][0]
+        tvec[0][0][1] = -tvec[0][0][1]
+        tvec[0][0][2] = -tvec[0][0][2]
 
-    rvec, tvec, frame, center, p_k = detect(image, camera_matrix, dist_coefs, chess_piece)
+        rvec[0][0][1] = -rvec[0][0][1]
+        rvec[0][0][2] = -rvec[0][0][2]
+        m = compositeArray(cv2.Rodrigues(rvec)[0], tvec[0][0])
+        glPushMatrix()
+        glLoadMatrixd(m.T)
 
-    draw_background(frame)
+        glTranslatef(0, 0, -0.5)
+        glCallList(chess_piece.gl_list)
+        glPopMatrix()
 
-    if len(rvec) != 0 and len(tvec) != 0:
-        draw_obj(rvec, tvec, center, p_k)
+    glPopMatrix()
 
+    glFlush();
     glutSwapBuffers()
 
 
-def reshape():
-    if height == 0:
-      height = 1
+def compositeArray(rvec, tvec):
+    v = np.c_[rvec, tvec.T]
+    #print(v)
+    v_ = np.r_[v, np.array([[0,0,0,1]])]
+    return v_
 
-    glViewport(0, 0, width, height)
-    glMatrixMode(GL_PROJECTION)
+def init():
+    global chess_piece, texture_background
+    #glClearColor(0.7, 0.7, 0.7, 0.7)
+    glClearColor(0.0, 0.0, 0.0, 1.0)
+    glEnable(GL_DEPTH_TEST)
+    glShadeModel(GL_SMOOTH)
+    glEnable(GL_LIGHTING)
+    glEnable(GL_LIGHT0)
 
-    glLoadIdentity()
+    chess_piece = OBJ(obj_path)
 
-    if width <= height:
-      glOrtho(-nRange, nRange, -nRange*h/w, nRange*h/w, -nRange, nRange)
-    else:
-      glOrtho(-nRange*w/h, nRange*w/h, -nRange, nRange, -nRange, nRange)
+    glEnable(GL_TEXTURE_2D)
+    texture_background = glGenTextures(1)
 
-    glMatrixMode(GL_MODELVIEW)
-    glLoadIdentity()
+def keyboard(key, x, y):
+    sys.exit()
 
+if __name__ == "__main__":
+    # glutInitWindowPosition(0, 0);
+    # glutInitWindowSize(windowWidth, windowHeight);
+    glutInit(sys.argv)
 
-############################### APPLICATION
-def init_application():
-    print("Initializing app")
-    glutInit()
-    glutInitDisplayMode(GLUT_DOUBLE | GLUT_RGBA | GLUT_DEPTH)
-    glutInitWindowSize(width, height)
-    glutInitWindowPosition(0, 0)
-    glutCreateWindow("AR Chess")
+    glutSetOption(GLUT_ACTION_ON_WINDOW_CLOSE, GLUT_ACTION_GLUTMAINLOOP_RETURNS);
+    glutInitDisplayMode(GLUT_RGBA | GLUT_DOUBLE | GLUT_DEPTH)
+    glutCreateWindow("AR CHESS")
+    glutDisplayFunc(draw)
+    glutKeyboardFunc(keyboard)
+    glutIdleFunc(draw)
 
-    glutDisplayFunc(display)
-    glutIdleFunc(display)
-    # glutReshapeFunc(reshape)
+    glutFullScreen()
+    init()
 
-    init_gl()
-    print("Inizialization completed")
     glutMainLoop()
-
-
-
-init_application()
