+++ utils/paths.py
@@ -2,7 +2,7 @@ test_path = "resources/marker/image_1.jpg"
 video_path = "resources/marker/videooh_2.mp4"
 model_path = "resources/marker/aruco/model_50.png"
 obj_path = "resources/chess_models/Alfiere_Bianco/12929_WoodenChessBishopSideA_v1_l3_t.obj"
-# obj_path = "resources/chess_models/Scacchiera/10586_Chess_Board_v2_Iterations-2.obj"
+# obj_path = "resources/chess_models/Alfiere_Bianco/10586_Chess_Board_v2_Iterations-2.obj"
 # obj_path = "resources/chess_models2/chess-horse-obj/chess-horse.obj"
 # texture_path = "resources/chess_models/Alfiere_Bianco/12929_WoodenChessBishopSideA_diffuse.jpg"
 texture_path = "resources/chess_models/Alfiere_Bianco/download.png"
